daemon: move settings `jsonrpc_` methods to an individual module
